refactor: 로또 번호 생성기 도메인 패키지로 이동
