+++ src/main/java/lotto/domain/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
