+++ src/test/java/lotto/domain/LottoNumberGeneratorTest.java
@@ -1,10 +1,11 @@
-package lotto.util;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import lotto.domain.LottoNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
